refactoring: Change LoggerSetting's name to Setting
